requirements.txt
tensorflow>=2.9.0
numpy>=1.22.0
matplotlib>=3.5.0
opencv-python>=4.5.0
flask>=2.0.0
pillow>=9.0.0
pdfkit>=1.0.0
reportlab>=4.0.0
scikit-learn>=1.0.0
werkzeug>=2.0.0
tensorflow-addons>=0.16.1
pandas>=1.3.5
scikit-image>=0.19.3
scipy>=1.8.0
seaborn>=0.11.2
tqdm>=4.64.0
imgaug>=0.4.0
efficientnet>=1.1.1
tensorboard>=2.8.0
pyyaml>=6.0
imutils>=0.5.4
jupyter>=1.0.0
albumentations>=1.0.3
joblib>=1.1.0 